services/auth-service/requirements.txt
Flask==2.2.5
PyJWT==2.8.0
gunicorn==20.1.0